backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
deepgram-sdk==3.2.7
python-dotenv==1.0.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

